Move shared classes and utils to /src/shared
