+++ models/patientDelivery.js
@@ -2,7 +2,7 @@ const _ = require('lodash');
 const { Model } = require('sequelize');
 const { DeliveryStatus } = require('../src/shared/constants');
 const metadata = require('../src/shared/metadata/patientDelivery');
-const convertToSequelizeField = require('../src/shared/metadata/convertToSequelizeField');
+const convertToSequelizeField = require('../src/shared/convertToSequelizeField');
 
 module.exports = (sequelize) => {
   class PatientDelivery extends Model {
