+++ src/shared/metadata/patientDelivery.js
@@ -1,4 +1,4 @@
-const ModelMetadata = require('./ModelMetadata');
+const ModelMetadata = require('../ModelMetadata');
 const { DeliveryStatus } = require('../constants');
 
 const fields = [
